+++ src/main/java/de/maxhenkel/gravestone/events/DeathEvents.java
@@ -37,6 +37,8 @@ public class DeathEvents {
             return;
         }
 
+        GravestoneMod.LOGGER.info("The death ID of player {} is {}", death.getPlayerName(), death.getId());
+
         BlockPos graveStoneLocation = GraveUtils.getGraveStoneLocation(world, death.getBlockPos());
 
         if (GravestoneMod.SERVER_CONFIG.giveObituaries.get()) {