+++ src/main/java/de/maxhenkel/gravestone/events/DeathEvents.java
@@ -93,7 +93,7 @@ public class DeathEvents {
 
     public static boolean keepInventory(Player player) {
         try {
-            MinecraftServer server = player.getServer();
+            MinecraftServer server = player.level().getServer();
             if (server == null) {
                 return false;
             }